services/notes-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.3.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
